Upgrade to celery 4.4.7
lazr.jobrunner declares that it requires celery >= 4.3 on Python 3.
Since our workers apparently still work with celery 4.1.1 on Python 3,
using a lower version may not in fact be a critical problem (I believe
the dependency was due to https://github.com/celery/celery/pull/5370,
which only matters when `CELERY_CONFIG_MODULE` points to a module within
a PEP 420 namespace package), but it doesn't seem like a good idea.