Fix position tracking in EncodableLibraryFileAlias
Python's `len()` requires that `__len__` returns a value greater than or
equal to 0.  `EncodableLibraryFileAlias` violated this in the case where
a caller passed an explicit length to its `read` method which was
greater than the number of bytes remaining.  In this case, increase the
position only up to the end of the file, not past it.

LP: #1940664