Authenticate using +launchpad-services rather than empty username
For private snap and OCI recipe builds, we previously authenticated to
git using the empty username and a non-user-bound macaroon.  This is
syntactically valid and works with current versions of git, but it's the
sort of edge case in the URL specification that's easy to mishandle.

curl 7.58.0 (used by git's HTTP transport in Ubuntu 18.04) fails to send
an Authorization header in this case, while curl 7.68.0 (used by git's
HTTP transport in Ubuntu 20.04) gets it right.  I bisected this to
https://github.com/curl/curl/commit/46e164069d, which is a substantial
refactoring of URL parsing that would be unreasonable to try to
backport.  Using a reserved username instead is safer.

LP: #1940804