Upgrade to python-keystoneclient 3.21.0
Now that the production librarian no longer needs to talk to PS4
(Icehouse), I'm not as scared of upgrading keystoneclient to a less
prehistoric version.  I chose the version from OpenStack Train to match
our current swiftclient version.

This means that our dependency tree is finally consistent enough for
"pip check" to pass.