Add explicit model for charm bases
This allows us to configure the build channels used for each base, and
the architectures for which builds for each base will be dispatched.  In
particular, it will allow us to avoid dispatching i386 charm recipe
builds for Ubuntu 20.04.

The model is essentially a cut-down version of `SnapBase`, looked up by
distroseries instead of name since that's how `charmcraft.yaml`
specifies bases.