Save a query in factory.makePerson
Since this method knows that it's just created the `Person`, it doesn't
need to perform the extra check for an existing preferred email address
in `Person.validateAndEnsurePreferredEmail` and can instead just call
`Person.setPreferredEmail` directly.  This saves a query from every
`makePerson` call, which is very common in the test suite.