Suppress unnecessary linecache check in tracebacks
While looking for something else in `strace` output, I happened to
notice lots of `stat` calls for Python source files.  It turns out that
the `traceback` module uses `linecache` to cache the contents of source
files, and it checks that the cache is up to date every time it
generates a traceback.  In typical Python code this isn't a big deal,
but Launchpad generates a lot of tracebacks: for instance, every query
while a request timeline is active causes a traceback to be stored in
the timeline.  This results in a lot of system call noise.

Since production deployments restart processes, and since gunicorn is
configured to auto-reload in development, it seems reasonably safe to
suppress these cache checks.  I don't have accurate measurements of how
much time it saves, but it certainly makes `strace` output less noisy.