copy-apache-config: Byte-compile using $(PYTHON), not $(PY)
`make copy-apache-config` (called from `make install`) has to
byte-compile `scripts/_pythonpath.py` to avoid problems with Apache
doing so as root.  However, Launchpad's virtualenv may not exist yet, as
for example when `make install` is being run by `rocketfuel-setup`, so
we need to use the system Python rather than the one in the virtualenv.
The Python version is the same either way, so this makes no difference
to byte-compilation.