Ensure that the sourcecode directory exists
Since the last file in `sourcecode/` was removed (in 27c3b86f17),
`rocketfuel-get` called from a `rocketfuel-setup` run no longer
automatically symlinks dependencies into the new Launchpad clone,
because it looks for directories with `sourcecode` subdirectories as
targets.  Add a harmless file there so that `git` creates the containing
directory.