Prevent Translation-en publishing collisions
If the series is configured to do so, the publisher may generate
Translation-en files with long descriptions extracted from Packages
files.  However, it's also possible for DDTP tarballs to include
Translation-en files, which then collide with the publisher when
extracted and produce either assertion failures or inconsistent Release
files depending on which one happened to win.  Filter these out.

I fixed a ResourceWarning in passing.

LP: #1900464