Simplify sets-of-suites handling in the publisher
Rather than storing sets of tuples of (distroseries name, pocket), it's
equivalent and simpler to just store sets of suite names, which are
composed from the distroseries name and the pocket.  This means that we
don't have to worry about details such as whether security-proxied
enumeration items behave identically to unproxied items for purposes of
set membership checks (in some situations they don't).  The code is also
slightly shorter and easier to follow this way.

I have no proof that this changes behaviour, although it's possible that
it will fix some weirdness that I encountered while trying to force the
jammy release pocket to republish to work around a separate bug on
2021-10-16; while debugging that, I noticed some oddities with set
membership checks that I never quite got to the bottom of.