Apply pyupgrade
I've started out with some conservative options that make minimal
changes, so at the moment most of the changes are for things like using
more modern set literal and dictionary comprehension syntax.  We can
apply further options to upgrade to Python 3 syntax as a separate step.

`contrib` is excluded (as with `flake8`) because it consists of code
mainly copied from elsewhere.