Update EnumCol doctest to test DBEnum instead
The tests needed are quite similar, but using the Storm API requires
some small tweaks to the mechanism.

In the process of doing this, I noticed that DBEnum only checks the type
of the `enum` parameter when the property is used (thus constructing the
associated variable), rather than when the property is constructed; and
it also failed to actually do the type check in practice due to a Python
3 porting bug.  This seems inconvenient, so I moved the type check
earlier and fixed it for Python 3.