Publish OOPSes to disk by default in development config
The development config previously wrote OOPSes to rabbitmq if it was
running (which it normally would be within `make run`).  But this is
quite inconvenient for normal development purposes: unless you're also
running something like python-oops-tools to receive OOPS reports over
AMQP, they effectively disappear into the void and are difficult to
find.

Although the test runner needs to preserve the previous behaviour, a
better default for interactive use in `LPCONFIG=development` is to
publish OOPSes to disk.