Allow access token targets to be containers
A test failure in
https://code.launchpad.net/~ilasc/launchpad/+git/launchpad/+merge/410373
revealed that it's cumbersome to require access token targets to match
the context object exactly: that requires all the interesting methods to
be implemented directly on the access token target, and in the case of
(for example) updating a revision status report on a commit in a Git
repository, that's quite cumbersome.

A better approach would be to work the same way as OAuth token scopes:
in those cases, an exact match is fine, but so is a match for the
nearest `LaunchpadContainer` in its URL chain, or one of its parent
containers.  This allows access tokens for Git repositories to be valid
for other objects that are subordinate to those repositories.