Add and honour DistroSeries.proposed_not_automatic
If set, this flag causes the -proposed pocket of the series to have
`NotAutomatic: yes` and `ButAutomaticUpgrades: yes` written to its
`Release` file, causing apt to not install versions from that suite
without explicit user consent, but to apply upgrades if the user has
already installed a version higher than that available from automatic
suites.

See "man apt_preferences" and
https://wiki.debian.org/DebianRepository/Format for detailed rules.
They're quite confusing, but this is the closest available approximation
of "only upgrade a package from -proposed if you already installed an
earlier version from -proposed".

This is analogous to, and largely copied from, the existing
`DistroSeries.backports_not_automatic` flag.

LP: #1016776