Cope with running outside a git tree
It would be useful to be able to produce Launchpad deployment artifacts
that have their version information pregenerated and don't require a git
working tree.  This only requires a small number of tweaks: the main one
is that we need to ensure that `talisker.config` has a usable strategy
for discovering the current revision, which is just a matter of writing
a `version-info.txt` file to go alongside `version-info.py`.