Rearrange Makefile phony target declarations
Phony targets in Makefiles are those that don't correspond to a real
file name.  Launchpad previously declared all these in one go at the end
of its Makefile, but this was hard to read and it was easy for them to
get out of sync (there were a few entries that didn't correspond to
actual targets any more).  Instead, move all the declarations to be
immediately before the target in question.

I went through all the existing targets and added phony declarations for
a few that were previously missing.  There were also a few existing
phony declarations for targets that do correspond to real file names
(`$(LP_JS_BUILD)`, `TAGS`, and `tags`): these appear to have been
deliberate in order to make sure that these targets are always rebuilt,
so I left those in place.