Support publishing deployment artifact to Swift
This adds `build-tarball` and `publish-tarball` make targets, which will
help us to write a Jenkins job that builds a deployment artifact and
publishes it to Swift for use by future deployment machinery.

The new `ols-vms.conf` file is to assist integration with
lp:ols-jenkaas, which is where we currently do publication to Swift for
other projects.

Aside from the new build and publish scripts, I had to adjust the
existing wheel-building target a bit to also build wheels for the
packages in `requirements/setup.txt`.