Update buildd-manager's open file descriptor limit
We added 30 new builders to Launchpad's production build farm today (for
a total of 348), after which we started seeing "Too many open files"
from buildd-manager.  This is because the value we set for
`RLIMIT_NOFILE` was tuned for the old in-process download strategy,
which we replaced in July 2020.  Re-tune it for the new
process-pool-based download strategy.