Adjust ObjectSet.__iter__ to avoid pyupgrade confusion
pyupgrade rewrites `six.itervalues(dct)` as `dct.values()`, which is
usually correct, but not when returned from an `__iter__` method.  That
must return an actual iterator, and returning a view object instead
raises an exception:

  TypeError: iter() returned non-iterator of type 'dict_values'

Fixing this in pyupgrade seems to be hard as the AST walk doesn't have
enough information about ancestor nodes beyond the immediate parent, so
just manually fix the one place in Launchpad where this is a problem so
that pyupgrade can handle other instances of this pattern mechanically.