Extend ppa-generate-keys to generate keys for copy archives too
It's currently awkward to run builds that use the output of test
rebuilds, since they're only signed if we go through some manual steps
to arrange that.  Since the copy archive publisher has access to the
signing service, having it generate signing keys is straightforward
enough, so let's just do that.

This requires a separate invocation with the `--copy-archives` option
(mainly since copy archives are published on a different machine so the
signing keys need to be authorized for use by that machine).

Once we have this running automatically on production, we should change
`Archive.can_be_published` to have a guard for copy archives similar to
the one it has for PPAs: it should only publish production copy archives
that have had a signing key generated for them.