Rename mock builder slaves to workers
Nomenclature is a bit difficult here: within Launchpad, we need to
distinguish between the `Builder` database model and builder machines
(previously "builder slaves").  Using "builder worker" for instances of
the latter is a bit clunky, but in most cases it ends up just being
"worker" and that seems clear enough.