Apply "pyupgrade --py3-plus" to most remaining files
There are a couple of non-production scripts in `utilities/` and their
dependencies that still require Python 2, as well as the code in
`lib/contrib/` that's copied from elsewhere pending replacement by
dependencies from PyPI and that's better to leave alone.  Everything
else that `pyupgrade` knows how to deal with now uses Python 3 syntax.