Rename conjoined master/slave bug tasks to primary/replica
Launchpad bugs have a peculiar feature called "conjoined tasks".  For
example, the tasks on a bug that's being primarily worked on in the
jammy series of Ubuntu but whose fix then needs to be backported to
focal and impish might look like this:

  base-files (Ubuntu)    Status tracked in Jammy
   -> Focal              Triaged        Critical
   -> Impish             Triaged        Critical
   -> Jammy              In Progress    Critical

In this case, the task on base-files (Ubuntu) is currently referred to
as a "conjoined slave", and the task on base-files (Ubuntu Jammy) is its
corresponding "conjoined master".  The metadata of the conjoined slave
cannot be updated independently; instead, changes are automatically
propagated from the conjoined master.

This terminology obviously doesn't fit with inclusive naming standards.
Rename "conjoined master" to "conjoined primary" and "conjoined slave"
to "conjoined replica", which I think also expresses the metadata
replication relationship a little more clearly.

I also considered renaming "conjoined master" to "conjoined
series-specific task" (or `conjoined_specific`) and "conjoined slave" to
"conjoined generic task" (or `conjoined_generic`).  That had the
advantage of making it clearer which way round the tasks go (which I
always find confusing and have to look up), but it's more cumbersome in
prose, and it makes the replication relationship less clear.