Tolerate git commits with invalid author/committer emails
`email.utils.formataddr` requires email addresses to be ASCII, per the
email RFCs, and raises `UnicodeEncodeError` if they aren't.  However, by
the time a commit gets this far, we need to do something with it (it may
be an import from a repository hosted somewhere else, in which case we
wouldn't even theoretically be able to reject the bad metadata on push).
Leave the author or committer unset in this case and ensure that we can
still render the commit properly when we need to.

LP: #1958281