Rename "Edit access tokens" to "Manage access tokens"
This is consistent with other "Manage ..." links on `GitRepository` used
for things like permissions.

I think it also makes a little more sense for the "Manage access tokens"
link to be next to "Manage permissions" and "View activity", which deal
with other kinds of access control, and in any event "Delete repository"
should stay in the last position, so I've reordered the links a bit.