Tighten up lpcraft parsing
This now matches the relevant parts of
https://git.launchpad.net/lpcraft/tree/lpcraft/config.py a little more
closely: the top-level "pipeline" and "jobs" keys are now required, as
are the "series" and "architectures" keys of each job definition, which
between them are the only things Launchpad will need to look at.  Each
element of the "pipeline" key is now normalized into a list.

The returned configuration object also now implements a minimal
`ILPCraftConfiguration` interface, making it easier to handle cases
where it's returned from a security-proxied method and thus ends up
being security-proxied itself.