Delay copy archive publishing until signing keys exist
This should be safe on production now that
`cronscripts/ppa-generate-keys.py --copy-archives` is cronned, and it
means that we don't end up in a situation where we publish an unsigned
copy archive first and then don't realize that it needs to be
republished to pick up the signing key.

I cheated slightly by overloading this on
`config.personalpackagearchive.require_signing_keys`, but it doesn't
seem worth the churn of adding another configuration option just for
this and arranging for it to be enabled only on production.