Avoid showing duplicate builder information on build pages
Build pages previously assumed (perhaps not very intentionally) that
`BuildFarmJob.builder` wasn't set until the build finished.  However,
following
https://code.launchpad.net/~cjwatson/launchpad/+git/launchpad/+merge/414070,
we now set `BuildFarmJob.builder` as soon as we handle any status
response from the builder running the build, which isn't quite as early
as it could be but it's pretty close.  This means that we've started
rendering the builder name twice for builds in the `BUILDING` state,
which is unsightly.

Simplify build pages so that they only consider `BuildFarmJob.builder`
regardless of the build state, which avoids sometimes rendering the
builder name twice.