Properly clear builder from BuildQueue.reset
Following
https://code.launchpad.net/~cjwatson/launchpad/+git/launchpad/+merge/414070,
the `builder` attribute of build farm jobs is now set while the build is
in progress (and indeed this commit sets it slightly earlier as well,
from `BuildQueue.markAsBuilding`, mainly to maximize the chance of
discovering problems in tests).  This has various advantages.

However, `BuildQueue.reset` can be called when a builder fails, upon
which the job is likely to be dispatched to a different builder.  It
cleared `BuildQueue.builder` as part of resetting the build queue entry,
but not the corresponding attributes of the `BuildFarmJob` or the
specific build, which caused an assertion failure later on.  Repair this
by clearing those attributes too.

LP: #1960132