Don't retry dep-wait builds for disabled DASes
If a `DistroArchSeries` is disabled, then we won't create any new builds
for it; this is used when decommissioning obsolete architectures.  In
this situation, it also doesn't make sense to automatically retry builds
in the dependency-wait state.

(There is currently one build for trusty powerpc that's occasionally
being retried for this reason, and I'd like it to stop.)