Add basic support for commercial subscriptions to distributions
These are used to control the ability to use privacy features.

There isn't much in the way of tests here yet, because writing those
needs privacy support in the `Distribution` model which is a rather more
involved affair, but this gets some of the simple bits out of the way.