Introduce View and LimitedView permissions on distributions
For now these are trivially public, so no functional change is expected
here.  That will change as soon as we put distribution privacy in place.

As with projects, only the bare minimum of attributes remain public;
attributes needed by people who can see some individual artifacts of the
distribution (such as those needed to render links to it) are under
`LimitedView`; and everything else that was formerly public is now under
`View`.