Use slightly shorter test DB instance names
Process IDs may nowadays be up to seven digits long, so the previous
test database instance names we used could cause
'launchpad_ftest_template_{test_instance}' to be up to 65 characters
long; but PostgreSQL identifiers are limited to 63 characters, so this
could cause the tests in
`lp.services.webapp.tests.test_dbpolicy.TestFastDowntimeRollout` to
fail.  Use a slightly shorter random suffix to compensate for this.