Batch queue items in process-accepted
`PackageUploadSet.getAll` (and hence `DistroSeries.getPackageUploads`)
returns a `DecoratedResultSet`.  These are usually used in contexts
where the result is batched by a view or by the webservice machinery,
and they have a property that's slightly surprising outside those
contexts: if you iterate over a decorated result set without first
slicing it, then its `pre_iter_hook` is called for *all* rows in the
result set.  This means that `process-accepted` can run out of memory if
the number of accepted queue items for a given archive and series is
very large, as happened recently for a test rebuild archive.

To fix this, process queue items in batches.  We need to take a bit of
care here because the act of processing queue items changes the result
of the query, but explicitly ordering by ID and adding a threshold
condition deals with that.  The particular implementation I chose here
processes items in ascending order by ID rather than descending order as
was implicitly used previously, but I think that's slightly better
anyway.