Refactor lp.archivepublisher.diskpool using pathlib
The Artifactory bindings are based on `pathlib`, and so it makes it a
little easier for my forthcoming Artifactory publication work to have a
more `pathlib`-style implementation of `DiskPool` and `DiskPoolEntry`.

I don't think it makes sense to go all-in on `pathlib` while we're stuck
on Python 3.5, because Python 3.6 improved much of the standard library
to accept path-like objects as paths rather than just (byte) strings,
and without that we have to add some awkward stringifications in several
places.  However, at this scale it doesn't work out too badly.

I also added a few type annotations, mainly to serve as documentation
for the types I was changing.  These are the first type annotations in
Launchpad, and I haven't even tried to run `mypy` to check them yet.
Similarly to the above, I don't think it makes sense to go all-in on
type annotations until we're on at least Python 3.6 and can use variable
annotations, but I saw no reason not to at least add these.