Rename cron.base-ppa to cron.base-ppa.sh
This causes https://github.com/pre-commit/identify to identify it as
`shell` rather than as just a generic text file, which will shortly be
useful for adding a `shellcheck` pre-commit hook.

I also removed the `#!` line from this file; since the file is
non-executable and only ever processed by sourcing it into other shell
scripts, there's no point in it having a `#!` line to tell the kernel
how to execute it.