Recreate many triggers to be column-specific
I happened to notice that adding columns to the `Archive` table took
much longer than it should, which seems to be because it has a trigger
on all `UPDATE`s even though it only cares about two of its columns.  On
closer inspection quite a number of our triggers have a similar
property.  PostgreSQL supports making triggers be column-specific on
`UPDATE`, so do that in cases where the trigger only cares about a small
number of columns.

The `mv_branch_*_update` trigger functions also check whether `id` has
changed and explicitly raise an exception if so, but I ignored that
because (a) it's silly and (b) any case where it might somehow happen
and matter will violate foreign key constraints anyway.