Avoid holding transactions after some buildd-manager scans
buildd-manager is good at committing its transaction when it's
intentionally made database changes during a builder scan, but it
doesn't end (commit or abort) its transaction when it hasn't made any
database changes.  The effect of this is that if the last builder
scanned in a cycle requires no database changes (perhaps because it's
idle with nothing to build, or manual, or something along those lines),
then buildd-manager can hold a transaction until the start of the next
scan cycle.  Due to things like
`BinaryPackageBuildSet.addCandidateSelectionCriteria`, this can lock out
schema changes to some non-trivial tables such as `Archive`, possibly
causing them to exceed fastdowntime windows.

To avoid this, use `TransactionFreeOperation` to enforce the requirement
that `WorkerScanner.scan` does not hold a transaction.