Fix dispatch of CI build jobs
`CIBuild.getConfiguration` calls out to an external service, and so
can't be called from buildd-manager's main thread because it would block
the Twisted reactor.  A better approach would be to store the relevant
bits of job configuration in the database, but a quick fix for now is to
defer this request to a thread, similar to the approach taken in
`lp.soyuz.adapters.archivedependencies._get_sources_list_for_dependencies`
for retrieving GPG keys from the keyserver.

To test this, I extended `GitHostingFixture` to detect the case where
the default timeout hasn't been set, which would cause
`GitHostingClient` to fail in a real environment.  I'd like to make the
fixture do this by default, but attempting that caused a number of other
(probably artificial) test failures which would need to be fixed first.