Move publishing queries from DistroSeries to PublishingSet
The existence of `DistroSeries.getSourcePackagePublishing` and
`DistroSeries.getBinaryPackagePublishing` was a historical anomaly, as
this sort of publishing-specific code should live in `lp.soyuz` or
`lp.archivepublisher`.  Move these to the more appropriate
`PublishingSet.getSourcesForPublishing` and
`PublishingSet.getBinariesForPublishing`.

As well as cleaning up some minor technical debt, this helps to prepare
for Artifactory publishing, where we're likely to need to be able to
query for publications in an archive without filtering by series.