Request CIBuilds on ref creation as well as update
Most of the things we do when Git refs are updated really do only make
sense for refs that previously existed (e.g. a snap recipe can't become
stale as a result of a ref being created, because it can't have pointed
to that ref to begin with).  However, creating CI builds for new refs
makes sense, so arrange for that to happen on both create and update
events.