Refactor DiskPool.addFile to take an IPackageReleaseFile
For publishing to Artifactory, `addFile` will need a little more context
than it currently has: as well as the file contents and hash, it'll need
a reference to the originating `SourcePackageRelease` or
`BinaryPackageRelease` in order to be able to calculate initial
properties for the new artifact.

In support of this, refactor `ISourcePackageReleaseFile` and
`IBinaryPackageFile` to inherit from a common `IPackageReleaseFile`
interface, and refactor `DiskPool.addFile` to take an object providing
that interface rather than taking the file contents and hash separately.