Fix update-pkgcache crash with disabled archives
If an archive is disabled while `update-pkgcache` is running, then it
can cause a crash, because `DistributionSourcePackageCache.updateAll`
and `DistroSeriesPackageCache.updateAll` previously returned None rather
than 0 in that case.

It's only possible to hit this bug if an archive is disabled between the
time when `PackageCacheUpdater.main` fetches the list of archives for a
given distribution and the time when
`PackageCacheUpdater.updateDistributionCache` or
`PackageCacheUpdater.updateDistroSeriesCache` is called for the
newly-disabled archive (the fetched list of archives excludes disabled
archives, but `update-pkgcache` may commit transactions between those
points and so may see updated database state).  To simulate this
situation, I had to restructure the script somewhat in order to be able
to write in-process tests.