Fix wsgi-archive-auth.py for Ubuntu 18.04
`scripts/wsgi-archive-auth.py` runs within Apache mod_wsgi, which is a
difficult environment for Launchpad code: it doesn't activate
Launchpad's virtualenv and can't even be told to disable the automatic
import of the `site` module (similar to Python's `-S` command-line
option), so we have to take various countermeasures before manually
activating the virtualenv so that we can import other parts of
Launchpad.

In bionic, we need a new countermeasure.  The system `zope.interface`
package (installed as a dependency of landscape-common) is built with
setuptools >= 31.0.0, which has new namespace package handling that
causes `sys.modules["zope"]` to exist and to point to the
system-installed package, even without importing `zope.interface`; this
then confuses Launchpad code that tries to import from `zope.interface`
later.  Arrange for the namespace package to be re-imported to avoid
this problem.