Fix git arguments passed to CI builds
`CIBuildBehaviour` was passing the commit ID to builders using the
fictitious `commit_sha1` argument.  Since this is just intended to be
passed to `git checkout`, it doesn't make sense to invent a new name for
it; it can just use the same `git_path` argument that's used for passing
branch and tag names to other build types.

This also requires
https://code.launchpad.net/~cjwatson/launchpad-buildd/+git/launchpad-buildd/+merge/418543
in order to work properly (though it can land first since CI builds are
currently broken due to this pair of issues anyway).