Store CIBuild stages in the DB for dispatch
buildd-manager previously had to fetch a CI build's configuration again
from git in order to calculate the `jobs` argument to dispatch, which
was poor design: we should instead store everything that buildd-manager
needs in the database so that it can dispatch the job immediately.

This is incompatible with previously requested builds: without DB
surgery, it won't be possible to dispatch those builds.  Since this
feature is still rarely used and hasn't yet been announced, I think
that's OK.