Update CIBuild results in buildd-manager
We previously handled all updates to the results of `CIBuild`s in
archiveuploader.  The main problem with this approach is that without
some other rearrangements it only does anything for successful builds,
but we want to update the corresponding `RevisionStatusReport` rows in
unsuccessful cases too.  It also means that report updates are delayed
for some time until the serialized upload task completes, which isn't
ideal.

Move some of this work back to `storeLogFromWorker` so that report
results are updated in buildd-manager.  We still defer uploading
artifacts to the librarian to archiveuploader, though we now store the
`jobs` dictionary returned by the worker in the database rather than
marshalling it via a `jobs.json` file on disk.