Add an Artifactory pool implementation
This will allow publishing files to Artifactory, using approximately the
same API used by the existing `lp.archivepublisher.diskpool` in order to
simplify the process of integrating it into the publisher as far as
possible.  (The actual publisher integration will come in later
branches.)

The main difference between local and Artifactory publishing methods is
that for Artifactory Launchpad doesn't generate indexes itself, but
rather it sets properties on the artifacts in Artifactory to instruct
Artifactory on how each artifact should be indexed.  This means that we
need to be able to take an existing artifact and figure out which
publications are relevant to it.  We use some custom properties
("launchpad.release-id" and "launchpad.source-name") to assist with
this, as well as a custom "launchpad.channel" property which will be
used for snap-store-style channel publications.

Testing this is interesting, since we can't realistically bring up an
actual test Artifactory instance.  Instead, I added a fixture which
fakes the relevant API responses, derived by observing the behaviour of
the third-party bindings we're using (I've never seen the actual
Artifactory implementation).  This is inevitably incomplete and of
course there's a risk that it might be out of sync with production
behaviour, but it should be reasonably close and it will be easy enough
to fix if it turns out to be wrong.

At the moment this only really implements publishing for Debian-format
packages, although I included some early code related to Python wheels
as well to illustrate approximately how some of that is likely to work.