Fix race in TestWorker.test_status_after_build
`TestWorker.test_status_after_build` sends a build command to an
external launchpad-buildd worker and inspects the resulting status.
However, the build it dispatches will quickly fail (because we don't
have quite a full builder environment here, and nor do we have full
control over the external worker's event loop), and so depending on
timing we can end up seeing the status as either `BUILDING` or
`WAITING`.  The important thing here is that a build command dispatches
a build, so accept either possibility.