Implement Artifactory publishing method
The arrangements for identifying the target Artifactory repository are
somewhat basic (they currently use `Archive.name`, which isn't
guaranteed to be unique across a distribution), and the publisher
currently needs to fetch all the currently-published artifacts as part
of figuring out which properties it needs to update.  However, aside
from those problems, this is a functional basic implementation: I've
been able to use this to import a subset of Ubuntu focal and republish
it to a test repository in Artifactory.