Retry LibrarianLayer._check_and_reset a few times
We relatively often see mysterious `process-returncode` errors in
buildbot runs that don't seem to correspond to a failed test on the same
worker.  On closer inspection of the subunit stream, these seem to be
due to `LibrarianLayer._check_and_reset` getting `ECONNRESET` when
trying to check whether the librarian is still up.  I'm not sure exactly
why this might be happening, but it seems reasonable to retry the
are-you-still-there request a few times on general principles to see if
that makes things more resilient.