Allow editing Artifactory-related Archive columns
`Archive.publishing_method` and `Archive.repository_format` control
publishing to Artifactory, so we need a way to manage them.  For now,
restrict editing these to people with `launchpad.Admin` permissions on
the archive (since archives must have names corresponding to Artifactory
repository names and be in a distribution with appropriate
configuration), and only allow editing these if no packages have yet
been published to the archive (similar to privacy switching).