Fix git authorization for CI builds in private distributions
CI builds in private distributions fail with `fatal: repository
'https://git.launchpad.net/...' not found`.  This is because they
authenticate using the special `+launchpad-services` user and a
macaroon, and in that mode methods of `GitAPI` run with an anonymous
principal and are expected to use `removeSecurityProxy` rather than
doing normal security checks (see the comment near the top of
`run_with_login`).

The lookup infrastructure in `lp.code.model.gitlookup` mostly doesn't do
much in the way of permission checks, normally relying on the returned
repository's security adapter to check access grants.  However, there
are some exceptions: if the repository is in a source package or an OCI
project inside a private pillar, then the pillar is security-proxied
during traversal and so it implicitly performs permission checks.

Allow passing `check_permissions=False` to the lookup infrastructure to
suppress these checks, and pass this when performing a lookup as
`+launchpad-services`.  This is safe because it's only possible to use
that user in conjunction with a macaroon issued for a specific
repository.