Simplify symlink checks in rocketfuel-setup
`shellcheck` complains of two places in `rocketfuel-setup`:

    SC2012 (info): Use find instead of ls to better handle
    non-alphanumeric filenames.

I don't think its proposed solution is best in this case, but it has a
point about the problem: parsing the output of `ls -l` is pretty clunky.
We know exactly what we expect these two symlinks to point to, so test
them directly using `readlink` instead.