Hardcode the list of "risks" in Canonical's stores
Canonical's stores (the snap store, Charmhub, etc.) use a vocabulary of
"risks" that form part of channel names.  When we initially set this up
it made sense to query the snap store to get hold of that vocabulary.
However, the list of risks has been stable for some time and is unlikely
to change without substantial notice, and having to query an external
service just in order to do things like splitting a channel string into
components is a significant amount of complexity; the endpoint we're
querying is also deprecated.

Replace all this with a simple hardcoded enumeration, in `lp.registry`
since nowadays this applies to multiple applications.