Fix Distribution:+edit crash with enabled restricted processors
The main archive of a distribution may have an enabled processor which
is restricted, such that the owners of that distribution aren't
necessarily authorized to enable or disable it; riscv64 is one such for
Ubuntu at the moment.  This shouldn't prevent distribution owners from
making other changes in the distribution's +edit view.

PPAs already had a similar validation tweak.

LP: #1974330