Add method to upload a CI build to an archive
This new `Archive.uploadCIBuild` method will be used to publish the
output of CI builds in archives that publish via Artifactory.  It's
somewhat similar to `Archive.copyPackage`, and I considered calling it
`copyCIBuild` instead, but I decided to go with "upload" since it's only
usable for the initial copy into an archive, not for copies between
archives after that.

There's a new `CIBuildUploadJob` which fetches the output files from the
librarian and works out how to turn them into `BinaryPackageRelease`s.
This is inevitably package-type-specific, and for the time being is only
implemented for Python wheels, using `pkginfo` and `wheel-filename`.