lp.answers.browser: Apply black
This puts the necessary structure in place to be able to apply this
progressively rather than having to change the whole codebase at once.
The required isort configuration is incompatible with unblackened files,
so we need to split that pre-commit hook and run it twice with different
configuration on different subsets of files.