Add a trailing newline to Person:+sshkeys
It's annoying for text endpoints to return output without a trailing
newline.  For instance, it meant that running `curl` on this endpoint
produced confusing output because the next shell prompt appeared
immediately after the last SSH key with no intervening line break.

(In fact, POSIX requires text files to have a terminating newline.  See
https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_403
and
https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_206.)